feat: 实现内容发布完整链路（MinIO上传 + Kafka异步审核）
后端：
- 新增 MinIO 文件上传服务（MinioConfig + MinioService）
- Content/ContentImage/ContentVideo 实体对齐真实表结构（item_info/image_info/video_info）
- postMapper 支持 item_id 手动生成（MAX+1）及状态更新
- PostController 改为 multipart/form-data 接收文件+元数据
- PostServiceImpl 重写：参数校验 → 生成ID → 插入item_info → 上传MinIO → 落库媒体表 → 发Kafka审核
- KafkaConsumerService 改为消费 content-review topic 并更新审核状态
- 修复 CORS：范围从 /api/** 扩大为 /**
- FeedServiceImpl/inboxMapper/outboxMapper 适配新字段名

前端：
- api/post.js 改为 FormData 一次提交（文件+元数据）
- Compose.vue 新增标题输入，自动判断 contentType，去掉单独上传步骤
- PostComposer.vue 新增标题字段适配新接口
- composables/usePost.js 移除 feedStore.addPost

+++ src/main/java/org/sdb/buzzfeed/mapper/inboxMapper.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Mapper
 public interface inboxMapper {
 //    插入用户收件箱
-    Boolean insertInbox(List<Long> activeFans, int contentId, LocalDateTime publishTime);
+    Boolean insertInbox(List<Long> activeFans, Long contentId, LocalDateTime publishTime);
 //    下拉获取feed流
     List<Content> downFeed(Long userId, LocalDateTime lastTime, Integer num);
